requirements.txt
asgiref==3.8.1
boto3==1.39.4
botocore==1.39.4
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==45.0.5
cssselect2==0.8.0
distlib==0.3.9
dj-database-url==3.0.1
Django==5.2
django-allauth==65.9.0
django-axes==8.0.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-formtools==2.5.1
django-js-asset==3.1.2
django-mptt==0.17.0
django-otp==1.6.1
django-paypal==2.1
django-phonenumber-field==8.1.0
django-ranged-response==0.2.0
django-ratelimit==4.1.0
django-recaptcha==4.1.0
django-redis==6.0.0
django-simple-captcha==0.6.2
django-storages==1.14.6
django-two-factor-auth==1.17.0
django-widget-tweaks==1.5.0
dotenv==0.9.9
filelock==3.18.0
fonttools==4.58.5
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
pypi==2.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-memcached==1.62
pytz==2025.2
qrcode==7.4.2
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.13.0
setuptools==80.1.0
six==1.17.0
sqlparse==0.5.3
stripe==12.0.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.30.0
weasyprint==65.1
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.9.0
zopfli==0.2.3.post1
